+++ src/source.ts
@@ -1,5 +1,5 @@
-import { Config } from "./config";
-import { Diagnostics } from "./diagnostics";
+import { Config } from "./config.js";
+import { Diagnostics } from "./diagnostics.js";
 
 export interface Source {
 	/**
@@ -55,10 +55,6 @@ export interface SourceJustifyKeysOptions {
 	 * by another file that uses the same prefix for some reason.
 	 */
 	readonly isReserved?: (key: string) => boolean;
-	/**
-	 * If true, keys not starting with the specified prefix are replaced.
-	 */
-	readonly enforcePrefix?: boolean;
 }
 
 export interface SourceJustifyKeysResult {